OY-251: Add logging and timer for indexing. Bind full organisaatio cache population to batch size.
